DFR-1303 - move about to submit callback code into submit callback code for FR_approveApplication
